refactor(hooks): move all hooks to same dir
